added intro screen
